--- backend/src/main/java/codesquad/team4/issuetracker/exception/FileUploadException.java
@@ -1,11 +1,0 @@
-package codesquad.team4.issuetracker.exception;
-
-public class FileUploadException extends RuntimeException {
-    public FileUploadException(String message) {
-        super(message);
-    }
-
-    public FileUploadException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
