--- backend/src/main/java/codesquad/team4/issuetracker/exception/IssueStatusUpdateException.java
@@ -1,8 +1,0 @@
-package codesquad.team4.issuetracker.exception;
-
-
-public class IssueStatusUpdateException extends RuntimeException {
-    public IssueStatusUpdateException(String message) {
-        super (message);
-    }
-}
