+++ backend/src/main/java/codesquad/team4/issuetracker/exception/badrequest/FileUploadException.java
@@ -1,0 +1,8 @@
+package codesquad.team4.issuetracker.exception.badrequest;
+
+public class FileUploadException extends InvalidRequestException {
+    public FileUploadException(String message) {
+        super(message);
+    }
+
+}
