+++ backend/src/main/java/codesquad/team4/issuetracker/exception/badrequest/InvalidRequestException.java
@@ -1,0 +1,8 @@
+package codesquad.team4.issuetracker.exception.badrequest;
+
+public abstract class InvalidRequestException extends RuntimeException {
+
+    public InvalidRequestException(String message) {
+        super(message);
+    }
+}
