+++ backend/src/main/java/codesquad/team4/issuetracker/exception/badrequest/IssueStatusUpdateException.java
@@ -1,0 +1,8 @@
+package codesquad.team4.issuetracker.exception.badrequest;
+
+
+public class IssueStatusUpdateException extends InvalidRequestException {
+    public IssueStatusUpdateException(String message) {
+        super (message);
+    }
+}
