+++ backend/src/main/java/codesquad/team4/issuetracker/exception/notfound/AssigneeNotFoundException.java
@@ -1,10 +1,10 @@
-package codesquad.team4.issuetracker.exception;
+package codesquad.team4.issuetracker.exception.notfound;
 
 import static codesquad.team4.issuetracker.exception.ExceptionMessage.NOT_FOUND_ASSIGNEE;
 
 import java.util.Set;
 
-public class AssigneeNotFoundException extends RuntimeException {
+public class AssigneeNotFoundException extends DataNotFoundException {
 
     public AssigneeNotFoundException(Set<Long> assigneeIds) {
         super(NOT_FOUND_ASSIGNEE + assigneeIds);
