+++ backend/src/main/java/codesquad/team4/issuetracker/exception/notfound/CommentNotFoundException.java
@@ -1,8 +1,8 @@
-package codesquad.team4.issuetracker.exception;
+package codesquad.team4.issuetracker.exception.notfound;
 
 import static codesquad.team4.issuetracker.exception.ExceptionMessage.NOT_FOUND_COMMENT;
 
-public class CommentNotFoundException extends RuntimeException {
+public class CommentNotFoundException extends DataNotFoundException {
     public CommentNotFoundException(Long commentId) {
         super(NOT_FOUND_COMMENT + commentId);
     }
