+++ backend/src/main/java/codesquad/team4/issuetracker/exception/notfound/DataNotFoundException.java
@@ -1,0 +1,8 @@
+package codesquad.team4.issuetracker.exception.notfound;
+
+public abstract class DataNotFoundException extends RuntimeException {
+
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+}
