+++ backend/src/main/java/codesquad/team4/issuetracker/exception/notfound/LabelNotFoundException.java
@@ -1,10 +1,10 @@
-package codesquad.team4.issuetracker.exception;
+package codesquad.team4.issuetracker.exception.notfound;
 
 import static codesquad.team4.issuetracker.exception.ExceptionMessage.NOT_FOUND_LABEL;
 
 import java.util.Set;
 
-public class LabelNotFoundException extends RuntimeException {
+public class LabelNotFoundException extends DataNotFoundException {
 
     public LabelNotFoundException(Set<Long> labelIds) {
         super(NOT_FOUND_LABEL + labelIds);
